Fix #64 compare issue
This seems to have been caused by what must be a subtle difference
in how html gets parsed initially vs. on insertion. Fixing the
open and close form tags so that they cleanly enclose elements
(as opposed to opening the form deep in the dom, and closing it
at the top level), seems to have fixed the issue.